Continue refactoring by removing namespace hacks
This starts to refactor the samplers to:

1. actually use super(self, type).__init__ to correctly call out to
AbstractSampler's init or Gibbs's init
2. Instead of doing the exec-based dictionary flattening to facilitate
memoizing, we will instead refer
to all memoized variables as st.X, where st = self.state for each sampler. This
will cause all mutations to be stored in a coherent namespace, be short enough to use,
but be flexible enough to extend. This involved building a namespace class, too.
3. add attendent helper methods, like __repr__ and __getitem__ to make stuff
feel more natural