dispatch to the correct form of sampling (precision or covariance)
This changes the samplers to use a precision or a covariance, depending
on which is faster. For the SMA samplers, using the covariance provides
a faster sparse log determinant, since the covariance
(I + lambda W)(I + lambda W)'
is sparse. For the SE samplers, using the precision provides a faster
sparse log determinant, since the precision is:
(I - lambda W)'(I - lambda W)
which is not as sparse as the SMA form, but is still quite sparse.